backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database (将来の拡張用)
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9

# AI/ML
google-generativeai==0.3.0

# Data Processing
requests==2.31.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
httpx==0.25.1
